+++ vclib/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/InMemoryIssuerCredentialStore.kt
@@ -1,12 +1,10 @@
 package at.asitplus.wallet.lib.agent
 
-import at.asitplus.wallet.lib.data.CredentialSubject
-import at.asitplus.wallet.lib.iso.IssuerSignedItem
+import at.asitplus.wallet.lib.CryptoPublicKey
 import at.asitplus.wallet.lib.iso.sha256
 import io.matthewnelson.encoding.base16.Base16
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.datetime.Instant
-
 
 class InMemoryIssuerCredentialStore : IssuerCredentialStore {
 
@@ -20,51 +18,23 @@ class InMemoryIssuerCredentialStore : IssuerCredentialStore {
     private val map = mutableMapOf<Int, MutableList<Credential>>()
 
     override fun storeGetNextIndex(
-        vcId: String,
-        credentialSubject: CredentialSubject,
+        credential: IssuerCredentialStore.Credential,
+        subjectPublicKey: CryptoPublicKey,
         issuanceDate: Instant,
         expirationDate: Instant,
         timePeriod: Int
     ): Long {
         val list = map.getOrPut(timePeriod) { mutableListOf() }
         val newIndex = (list.maxOfOrNull { it.statusListIndex } ?: 0) + 1
+        val vcId = when (credential) {
+            is IssuerCredentialStore.Credential.Iso -> credential.issuerSignedItemList.toString().encodeToByteArray()
+                .sha256().encodeToString(Base16(strict = true))
+
+            is IssuerCredentialStore.Credential.VcJwt -> credential.vcId
+            is IssuerCredentialStore.Credential.VcSd -> credential.vcId
+        }
         list += Credential(
             vcId = vcId,
-            statusListIndex = newIndex,
-            revoked = false,
-            expirationDate = expirationDate
-        )
-        return newIndex
-    }
-
-    override fun storeGetNextIndex(
-        vcId: String,
-        subjectId: String,
-        issuanceDate: Instant,
-        expirationDate: Instant,
-        timePeriod: Int
-    ): Long {
-        val list = map.getOrPut(timePeriod) { mutableListOf() }
-        val newIndex = (list.maxOfOrNull { it.statusListIndex } ?: 0) + 1
-        list += Credential(
-            vcId = vcId,
-            statusListIndex = newIndex,
-            revoked = false,
-            expirationDate = expirationDate
-        )
-        return newIndex
-    }
-
-    override fun storeGetNextIndex(
-        issuerSignedItemList: List<IssuerSignedItem>,
-        issuanceDate: Instant,
-        expirationDate: Instant,
-        timePeriod: Int
-    ): Long {
-        val list = map.getOrPut(timePeriod) { mutableListOf() }
-        val newIndex = (list.maxOfOrNull { it.statusListIndex } ?: 0) + 1
-        list += Credential(
-            vcId = issuerSignedItemList.toString().encodeToByteArray().sha256().encodeToString(Base16(strict = true)),
             statusListIndex = newIndex,
             revoked = false,
             expirationDate = expirationDate
@@ -73,11 +43,11 @@ class InMemoryIssuerCredentialStore : IssuerCredentialStore {
     }
 
     override fun getRevokedStatusListIndexList(timePeriod: Int): Collection<Long> {
-        return  map.getOrPut(timePeriod) { mutableListOf() }.filter { it.revoked }.map { it.statusListIndex }
+        return map.getOrPut(timePeriod) { mutableListOf() }.filter { it.revoked }.map { it.statusListIndex }
     }
 
     override fun revoke(vcId: String, timePeriod: Int): Boolean {
-        val entry =  map.getOrPut(timePeriod) { mutableListOf() }.find { it.vcId == vcId } ?: return false
+        val entry = map.getOrPut(timePeriod) { mutableListOf() }.find { it.vcId == vcId } ?: return false
         entry.revoked = true
         return true
     }